backend/requirements/base.txt
django-environ==0.7.0
Django==3.2.7
django-cors-headers==3.10.0
django-filter==21.1
djangorestframework-simplejwt==4.8.0
djangorestframework==3.12.4
psycopg2==2.9.1
PyJWT==2.1.0
pytz==2021.1